Update with-stitches example
