+++ examples/with-stitches/components/StitchesLogo.jsx
